test(trip-direction): Created tests for helper `deep-object-clone`
