--- cheapTrip/src/app/trip-direction/helpers/trip-direction-helpers.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class TripDirectionHelpersService {
-  deepObjectClone<T extends Object>(value: T): T {
-    return JSON.parse(JSON.stringify(value));
-  }
-}
